requirements.txt
requests>=2.31.0
clickhouse-connect>=0.7.0
python-dotenv>=1.0.0
schedule>=1.2.0
